finetuning/requirements.txt
matplotlib==3.10.6
numpy==1.26.4
Pillow==10.4.0
qwen_vl_utils==0.0.14
transformers==4.51.3
vllm==0.8.2
accelerate==1.10.1
flash-attn==2.7.4.post1
gradio==4.44.1
gradio_image_prompter==0.1.0
pydantic==2.10.6
pycocotools==2.0.10
shapely==2.1.2
gradio_bbox_annotator==0.1.1
mmengine==0.10.7
deepspeed==0.16.7
ujson==5.11.0
wandb==0.22.2
liger-kernel==0.6.2
mpi4py==4.1.1
datasets==4.3.0
omegaconf==2.3.0
tensordict==0.10.0
codetiming==1.4.0
torchdata==0.11.0
tensorboard==2.20.0
ray[default]